refactor(saves): extract StateService class to saves/state.py
StateService owns the save-sync state lifecycle: in-memory dict,
on-disk JSON path, fcntl lock, default-state construction, schema
migrations (incl. _migrate_loaded_state, _migrate_saves_entries,
_strip_legacy_settings), load_state, save_state, prune_orphaned_state.

SaveService now composes self._state_svc = StateService(...) in its
ctor and aliases self._save_sync_state = self._state_svc.data so the
existing dict-access call sites elsewhere in SaveService continue to
work unchanged. Public lifecycle methods on SaveService remain as thin
delegates where external callers (tests, bootstrap) depend on them.

Method names preserved verbatim — naming hygiene (e.g. dropping the
redundant _state suffix inside StateService) is reserved for the
naming-consistency commit later in this PR.

Refs #257.
